llm_backend/requirements.txt
streamlit
langchain
langchain_community
langchain_google_genai
faiss-cpu
pandas
python-dotenv
fitz
frontend
backend
tools
